Add files via upload
Looking for the complete STL files of all the Printrbot Simple 1405 wood pieces? This is it!
Printrbot released the lasercut DXF drawing file, but we need STL to be able to print any of those parts. Nothing can convert non-water-tight drawings into printable stl, so they have to be drawn from scratch. (trace programs didn't work well either.) So I bought a copy of CorelDraw and learned hands on to draw with it. Since I was drawing a couple parts, I decided to go ahead and draw every single part that is usually made of wood. A hard week well spent..Finally done!
The original DXF can be found at https://www.youmagine.com/designs/printrbot-simple-maker-edition-1405
They are drawn specifically for printing. All the tabs are drawn slightly smaller than the holes they fit into. It is a nice, tight fit and first layer squish may have to be filed off to fit. Screw and nuts are drawn for the original 3mm hardware to be used.
I have no problem fitting parts that are printed at 2 perimeters, 0.